pyproject.toml
[project]
name = "job-offer-gen"
version = "0.1.0"
description = "job offer generator Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpipe>=4.50.0",
    "openpipe-art[skypilot]>=0.4.3",
    "python-dotenv>=1.0.1",
    "openai>=1.74.0",
    "boto3>=1.34.0",
    "async-lru>=2.0.5",
    "pydantic>=2.6.0",
    "wandb>=0.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src/job-offer"]
dev-mode-dirs = ["src"]


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/job-offer"]

[dependency-groups]
dev = [
    "semver>=3.0.4",
    "python-dotenv>=1.1.0",
    "openai>=1.74.0",
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "hatch>=1.14.1",
    "async-lru>=2.0.5",
    "accelerate>=1.7.0",
]
